added super basic planets

+++ Unity project/Solar Savings/Assets/Scripts/MoneyLogic/Data/Transaction.cs
@@ -4,21 +4,21 @@ using UnityEngine;
 
 public class Transaction
 {
-    public int amount { get; set; }
+    public float amount { get; set; }
     public System.DateTime time;
 
-    public Transaction(int inputAmount, System.DateTime inputTime)
+    public Transaction(float inputAmount, System.DateTime inputTime)
     {
         amount = inputAmount;
         time = inputTime;
     }
 
-    public Transaction(int inputAmount){
+    public Transaction(float inputAmount){
         amount = inputAmount;
         time = System.DateTime.Now;
             }
 
-    public void setValues(int inputAmount)
+    public void setValues(float inputAmount)
     {
         amount = inputAmount;
         time = System.DateTime.Now;
