+++ Unity project/Solar Savings/Assets/Scripts/MoneyLogic/Data/UserData.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class UserData
 {
-    public int totalBalance;
+    public float totalBalance;
 
     //public List<(int, System.DateTime)> transactionHistory = new List<(int, System.DateTime)>();
     public List<Transaction> transactionHistory = new List<Transaction>();
